src/db/streamdata/requirements.txt
mysql-connector-python==8.0.26
pandas